form input fixed - needs yup fix
