--- plants/src/components/AddPlantForm/data/apiURL.js
@@ -1,2 +1,0 @@
-// Remember your why
-export const apiURL = 'https://reqres.in/';
