--- plants/src/components/AddPlantForm/data/initialFormErrors.js
@@ -1,9 +1,0 @@
-// Remember your why
-
-const initialFormErrors = {
-    nickname: "",
-    species: "",
-    h2oFrequency: "",
-};
-
-export default initialFormErrors;
