--- plants/src/components/AddPlantForm/data/initialFormValues.js
@@ -1,8 +1,0 @@
-// Remember your why
-const initialFormValues = {
-    nickname: "",
-    species: "",
-    h2oFrequency: "",
-};
-
-export default  initialFormValues;
